+++ ilyvion.Laboratory/Extensions/EnumerableExtensions.cs
@@ -41,4 +41,19 @@ public static class EnumerableExtensions
         }
         return (num1, num2);
     }
+
+    public static (int min, int max) MinAndMax(this IEnumerable<int> source)
+    {
+        var (min, max) = source.Aggregate(
+            ((int?)null, (int?)null),
+            (acc, value) => (
+                acc.Item1.HasValue
+                    ? Mathf.Min(acc.Item1.Value, value)
+                    : value,
+                acc.Item2.HasValue
+                    ? Mathf.Max(acc.Item2.Value, value)
+                    : value));
+
+        return (min ?? 0, max ?? 0);
+    }
 }